requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
matplotlib>=3.5.0
pillow>=9.0.0


